% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_simplify.r
\name{cell.simplify}
\alias{cell.simplify}
\title{Get the simplified cell shape from junction points}
\usage{
cell.simplify(cell, cell.junctions)
}
\arguments{
\item{cell}{SpatialPolygon of the original cell}

\item{cell.junctions}{sp::SpatialPoints object containing all junction points
between cells. Can be obtained from the Analyse Skeleton function in ImageJ.}
}
\value{
An ordered matrix containing the x and y coordinates defining the simplified
cell.
}
\description{
Get the simplified cell shape from junction points
}
\details{
The returned matrix has n+1 rows (the first and last rows are the
same, so points can be joined up to form a closed polygon).
}
