% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cell_distances.r
\name{cell_cell_distances}
\alias{cell_cell_distances}
\title{Find the mean distance to k nearest neighbours}
\usage{
cell_cell_distances(cells, k)
}
\arguments{
\item{cells}{\code{sf} object.}

\item{k}{Numeric. The k nearest neighbours to calculate distances to.}
}
\value{
If \code{k} is a single value, returns a vector of same length as
\code{cells}, containing the distance (in pixels) between each cell and its
kth nearest neighbour. If \code{k} is a vector, returns a matrix with
\code{length(k)} variables, with the distance (in pixels) from each
cell to its nearest k neighbours.
}
\description{
Find the mean distance to k nearest neighbours
}
\details{
This code can be very slow if \code{cells} contains many polygons
or \code{k} has more than two values.
}
