% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_simplify.r
\name{cell_simplify}
\alias{cell_simplify}
\title{Get the simplified cell shape from junction points}
\usage{
cell_simplify(cell, cell.junctions, snap.tolerance)
}
\arguments{
\item{cell}{Obect of class \code{sf}; the original cell}

\item{cell.junctions}{Dataframe with two columns containing all junction points
between cells.}

\item{snap.tolerance}{Numeric. Value to buffer each cell to identify junction points.}
}
\value{
An ordered matrix containing the x and y coordinates defining the simplified
cell.
}
\description{
Get the simplified cell shape from junction points
}
\details{
The returned matrix has n+1 rows (the first and last rows are the
same, so points can be joined up to form a closed polygon).
}
