% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endwall_angles.r
\name{endwall_angles}
\alias{endwall_angles}
\title{Find the endwall angles of cells}
\usage{
endwall_angles(junction_points)
}
\arguments{
\item{junction_points}{This should be a matrix of the junction points defining
the polygon. The first and last points are the same (i.e. the line formed by
the points is closed).}
}
\value{
A list with the following components:
 \describe{
\item{slope1}{The slope of the leftmost endwall. Useful for plotting. }
\item{slope2}{The slope of the rightmost endwall.}
\item{angle1}{The angle of the leftmost endwall. Useful for analysis. }
\item{angle2}{The angle of the rightmost endwall.}
\item{points1}{The points defining the leftmost endwall.}
\item{points2}{The points defining the rightmost endwall.}
}
}
\description{
For a set of junction points that describe a simplified cell, this function
extracts the left-most wall and the right-most wall and calculates their
slope and bearing (in radians).
}
\details{
The endwalls are defined as the cell walls that define the length of
the cell along the leaf axis (i.e. stomatal north). Note that this trait is only
truly applicable to species that have broadly rectangular or polygonal cells
AND where the leaf axis is known (or can be identified from the image, i.e.
stomatal north). The \code{angle} outputs have been transformed to be within
-90 and 90, so that the distribution is centred on 0.
}
