% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidermal_trait_reliability.R
\name{epidermal_trait_reliability}
\alias{epidermal_trait_reliability}
\title{Estimate trait reliability}
\usage{
epidermal_trait_reliability(
  x,
  grouping.variable,
  heatmap = FALSE,
  thresholdval = 0.25
)
}
\arguments{
\item{x}{A data.frame with the trait measurements and grouping variable.
Names of traits should be the same as those produced by \code{extract_epidermal_traits()}}

\item{grouping.variable}{Character. Name of the column used to group images (e.g.
individual, species)}

\item{heatmap}{Logical. Plot a heatmap of trait reliability? Defaults to FALSE.}

\item{thresholdval}{Numeric. Used for heatmap plotting only - what is the
threshold for an 'unreliable' trait? Defaults to 0.25 (i.e. where the
variation in one group exceeds 25 percent of the variation in the entire imageset)}
}
\value{
Either a list or data.frame, depending on the value of \code{heatmap}. If
\code{heatmap = TRUE}, then a list of length 2 is returned, with a data.frame and
the ggplot heatmap object. If \code{heatmap = FALSE}, only the data.frame is
returned
}
\description{
Estimate trait reliability
}
\details{
This function generates estimates of intra-species or intra-
individual variability, expressed as a proportion of the variation in the
entire imageset.
}
\examples{
df <- epidermalmorph::podocarps
epidermal_trait_reliability(df, grouping.variable="individual")

}
