% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_image_features_options.r
\name{extract_image_features_options}
\alias{extract_image_features_options}
\title{Extract cell measurements from an image}
\usage{
extract_image_features_options(
  image.file,
  wall.skeleton.file,
  dir = ".",
  stomata.present = TRUE,
  stomatal.shape = TRUE,
  stomatal.arrangement = TRUE,
  pavement.cells = TRUE,
  sd.measures = TRUE,
  specific.inclusions = NULL,
  specific.exclusions = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image.file}{Character. File name of image.}

\item{wall.skeleton.file}{Character. File name of corresponding cell wall
skeleton image (see Details).}

\item{dir}{Character. Directory for images. Defaults to NULL.}

\item{stomata.present}{Logical. Does the image contain stomata? Defaults
to TRUE, see details.}

\item{stomatal.shape}{Logical. Include measurements of stomatal shape?
Defaults to TRUE, see details.}

\item{stomatal.arrangement}{Logical. Include measurements of stomatal
arrangement? Defaults to TRUE, see details.}

\item{pavement.cells}{Logical. Measure pavement cells? Defaults to TRUE,
see details.}

\item{sd.measures}{Logical. Return standard deviation of variables?
Defaults to TRUE, see details.}

\item{specific.inclusions}{Character. Cell measurements to include;
see details for options.}

\item{specific.exclusions}{Character. Cell measurements to exclude;
see details for options.}

\item{verbose}{Logical. Should progress output be printed in the console?
Defaults to FALSE.}
}
\value{
A list of length three, containing a vector with whole-image values, a
data.frame of pavement cell values and a data.frame of stomatal values.
}
\description{
Extract cell measurements from an image
}
\details{
This function contains the full pipeline from image to measurements.
The labelled image is required, along with the cell wall skeleton (with
junction points labelled), such as that produced by ImageJ's 'AnalyzeSkeleton'
function. If only a subset of variables is desired, this can can be set using the
arguments \code{ stomatal.shape, stomatal.arrangement, pavement.cells, sd.measures,
specific.inclusions,  specific.exclusions}. For a full list of variables and
graphical descriptions of what they measure, see the vignette here. ((add link))
}
