% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_edge_cells.r
\name{find_edge_cells}
\alias{find_edge_cells}
\title{Find edges of image, identify which cells are on the edge and which are}
\usage{
find_edge_cells(dir, fname, cells, rotated = NULL)
}
\arguments{
\item{dir}{String. Directory for images.}

\item{fname}{String. File name.}

\item{cells}{SpatialPolygonsDataFrame or list of Spatial PolygonsDataFrames.
See Details.}

\item{rotated}{Numeric. The angle that the image has been rotated. Defaults to NULL.}
}
\value{
A list with two components (see below). If \code{cells} is a list,
returns a list of lists.
#' \describe{
\item{edge}{Numeric. Vector containing the indices of all edge cells.}
\item{lose}{Numeric. Vector containing the indices of all cells on the bottom or right-hand edge.}
}
}
\description{
Find edges of image, identify which cells are on the edge and which are
}
\details{
The \code{cells} parameter can be the output from
\code{image_to_poly.R}, or another list of SpatialPolygons. Note: if
\code{cells} is a list, all items in the list should come from the same
image (i.e. different cell types).
}
