% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_junctions.r
\name{get_junctions}
\alias{get_junctions}
\title{Get junction point coordinates from skeleton image}
\usage{
get_junctions(wj)
}
\arguments{
\item{wj}{Raster of cell wall skeleton.}
}
\value{
Some SpatialPoints
}
\description{
Get junction point coordinates from skeleton image
}
