% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_connected_region.r
\name{grow_connected_region}
\alias{grow_connected_region}
\title{Grow a patch of cells}
\usage{
grow_connected_region(
  cells,
  nb,
  seed = NULL,
  existing.patch = NULL,
  blacklisted_cells
)
}
\arguments{
\item{cells}{SpatialPolygonsDataFrame of all cells, with edge cells removed.}

\item{nb}{Object of class "nb", created from the 'spdep' function \code{poly2nb()}.}

\item{seed}{Numeric. Index of first cell in patch (the 'seed' cell). Defaults to NULL.}

\item{existing.patch}{Numeric. Indices of cells in the patch which is to
be 'grown'. Defaults to NULL.}

\item{blacklisted_cells}{Numeric. Indices of cells to exclude (e.g. the subsidiary
cell of a stomatal complex that is cut off by the image edge).}
}
\value{
A numeric vector containing the indices of the sampled polygons
}
\description{
Grow a patch of cells
}
