% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_stomatal_complexes.r
\name{identify_stomatal_complexes}
\alias{identify_stomatal_complexes}
\title{Identify which cells belong to particular stomatal complexes}
\usage{
identify_stomatal_complexes(
  cells,
  paired = F,
  buffervalue,
  stom.val,
  subs.val = NULL
)
}
\arguments{
\item{cells}{Object of class \code{sf}. The image to be measured, converted
to polygons (e.g. using \code{image_to_poly()})}

\item{paired}{Logical. Are guard cells annotated separately (TRUE), or as one
cell (FALSE? Defaults to FALSE.}

\item{buffervalue}{Amount to buffer cells by. Should be similar to wall width (which varies with image scaling).}

\item{stom.val}{Numeric. Value of stomata in inpute image.}

\item{subs.val}{Numeric. Value of subsidiary cells in image (optional).
Defaults to NULL.}
}
\value{
sf object with the IDs of all cells belonging to the stomatal complex.
}
\description{
Identify which cells belong to particular stomatal complexes
}
