% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_poly.r
\name{image_to_poly}
\alias{image_to_poly}
\title{Convert image to SpatialPolygons}
\usage{
image_to_poly(dir, fname)
}
\arguments{
\item{dir}{String. Directory for images.}

\item{fname}{String. File name.

#' @return A list with components:
\describe{
\item{pave}{SpatialPolygonsDataFrame. Pavement cells. }
\item{stom}{SpatialPolygonsDataFrame. Stomata (pore + guard cells).}
\item{subs}{SpatialPolygonsDataFrame. Lateral subsidiary cells. }
}}
}
\value{
List of SpatialPolygonsDataFrames: pavement cells, subsidiary cells and stomata.
}
\description{
Convert image to SpatialPolygons
}
\details{
This function assumes that the pixels are coloured in the following order
(darkest to lightest): wall, stomate, subsidiary, pavement.
}
