% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMaxima.r
\name{localMaxima}
\alias{localMaxima}
\title{Find the localMaxima in a series}
\usage{
localMaxima(x)
}
\arguments{
\item{x}{Numeric}
}
\value{
A numeric vector with the positions of localMaxima
}
\description{
Find the localMaxima in a series
}
\details{
This function comes from user Tommy 's answer on SO:
https://stackoverflow.com/questions/6836409/finding-local-maxima-and-minima
}
