% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_amplitude.r
\name{max_amplitude}
\alias{max_amplitude}
\title{Find the maximum amplitude of cell sinuosity}
\usage{
max_amplitude(cell, junction_points)
}
\arguments{
\item{cell}{SpatialPolygon. The cell.}

\item{junction_points}{Matrix. The x and y coordinates defining the simplified
cell, as returned by \code{cell.simplify()}}
}
\value{
Numeric. The maximum distance from the simplified cell edge to the
actual cell wall.
}
\description{
Find the maximum amplitude of cell sinuosity
}
