% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncolour_functions.R
\name{nblue}
\alias{nblue}
\title{Generate 1:4 colours of in a particular colour family}
\usage{
nblue(x)
}
\arguments{
\item{x}{Number of colours (must be between 1 and 4)}
}
\value{
Character vector with colour names as hex values
}
\description{
Generate 1:4 colours of in a particular colour family
}
