% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_rows_stomata.r
\name{number_rows_stomata}
\alias{number_rows_stomata}
\title{Find the number of stomatal rows}
\usage{
number_rows_stomata(cells)
}
\arguments{
\item{cells}{SpatialPolygonsDataFrame, containing all cell polygons (i.e. not
just the stomatal polygons). The specific colour of the stomata is irrelevant,
but they should be the second-lowest value.}
}
\value{
A list with components:
\describe{
\item{n.rows}{Numeric. The number of rows of stomata}
\item{row.bounds}{Numeric. The y coordinates defining the edge of each row}
\item{row.wiggliness}{Numeric. The width of the stomatal band; see Details}
\item{row.consistency}{Numeric. The number of stomata per row.}
\item{stomrow.density}{Numeric. The number of rows of stomata per unit
length (per pixel if unscaled)}
}
}
\description{
Find the number of stomatal rows
}
\details{
This function assumes that the pixels are coloured in the following order
(darkest to lightest): wall, stomate, subsidiary, pavement. To calculate
\code{row.wiggliness}, principal components analysis was computed for the
stomatal pixels in each row. The value \code{row.wiggliness} is the range
of PC2; it is recommended that this value be standardised based on stomatal
 complex width, to account for stomatal size variability.

 @import stats
}
