% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_stomatal_complexes.r
\name{other_cluster_size}
\alias{other_cluster_size}
\title{Identify clustering in cells of type "other"}
\usage{
other_cluster_size(cells, val, buffervalue)
}
\arguments{
\item{cells}{Object of class \code{sf}. The image to be measured, converted
to polygons (e.g. using \code{image_to_poly()})}

\item{val}{Numeric. Value of cell typ eof interest.}

\item{buffervalue}{Amount to buffer cells by. Should be similar to wall width (which varies with image scaling).}
}
\value{
Numeric with size of median cluster (i.e. these cells are usually found in groups of X).
}
\description{
Identify clustering in cells of type "other"
}
