% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_subsample.r
\name{patch_subsample}
\alias{patch_subsample}
\title{Sample a connected patch of cells within an image}
\usage{
patch_subsample(
  cells,
  k.cells = NULL,
  k.pavement = NULL,
  k.stomata = NULL,
  plot.patch = FALSE,
  seed = NULL,
  stom.val = 85,
  subs.val = 170,
  pave.val = 255
)
}
\arguments{
\item{cells}{SpatialPolygonsDataFrame of all cells, with edge cells removed.}

\item{k.cells}{Numeric. Desired number of cells to be sampled in patch}

\item{k.pavement}{Numeric. Desired number of pavement cells to be sampled in patch}

\item{k.stomata}{Numeric. Desired number of stomata to be sampled in patch}

\item{plot.patch}{Logical. Plot the sampling iterations? Defaults to FALSE.}

\item{seed}{Numeric. For reproducibility, the starting cell can be set using
this argument.}

\item{stom.val}{Numeric vector of length 1. The value of stomatal cells (guard cells + pore)}

\item{subs.val}{Numeric vector of length 1. The value of subsidiary cells}

\item{pave.val}{Numeric vector of length 1. The value of pavement cells}
}
\value{
A numeric vector containing the indices of the sampled polygons
}
\description{
Sample a connected patch of cells within an image
}
\details{
If setting a seed using the \code{seed} argument, cell numbers can be
displayed on a plot of the cells using: \code{text(cells, labels = 1:length(cells))}.
}
\examples{
\dontrun{ #
#minimum patch size of 40 cells
patch_subsample(cells, k.cells=40)
#patch must include at least 10 stomatal complexes and 30 pavement cells
patch_subsample(cells,k.pavement=30, k.stomata=10)
}
}
