% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{podocarps}
\alias{podocarps}
\title{Epidermal traits example dataset}
\format{
A data frame with 9 rows and 57 variables.
}
\usage{
podocarps
}
\description{
A dataset containing the epidermal traits from three podocarp plants (three images per plant)
}
\keyword{datasets}
