% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_path_to_cell_type.r
\name{shortest_path_to_cell_type}
\alias{shortest_path_to_cell_type}
\title{Find the length of the shortest path to any cell of a particular type}
\usage{
shortest_path_to_cell_type(cells, path.matrix, cell.type.value)
}
\arguments{
\item{cells}{\code{sf} object.}

\item{path.matrix}{Matrix with path lengths between each pair of cells.
Recommended to use output from \code{cell_graph_shortest_paths()}.}

\item{cell.type.value}{Numeric. The value of stomata.}
}
\value{
A numeric vector with the distance (in cells) to the nearest cell
with value \code{cell.type.value}.
}
\description{
Find the length of the shortest path to any cell of a particular type
}
\details{
For cells with the same value as \code{cell.type.value}, this
 algorithm finds the path length to the nearest other cell of that type.
}
