% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_epidermal_traits.R
\name{subsample_epidermal_traits}
\alias{subsample_epidermal_traits}
\title{Measure epidermal traits from subsets of cells}
\usage{
subsample_epidermal_traits(
  x,
  n.runs = 100,
  k.cells = c(50),
  k.pavement = NULL,
  k.stomata = NULL,
  scaling = NULL,
  trait.subset = NULL,
  stom.val = 85,
  subs.val = 170,
  pave.val = 255,
  verbose = TRUE
)
}
\arguments{
\item{x}{x List of the form output by \code{extract_epidermal_traits()} (see documentation of that function for details).}

\item{n.runs}{Numeric. Number of times to sample each value of k.}

\item{k.cells}{Numeric. Stopping threshold for number of cells.}

\item{k.pavement}{Numeric. Stopping threshold for number of cells.}

\item{k.stomata}{Numeric. Stopping threshold for number of cells.}

\item{scaling}{Numeric. Global standard deviation values to standardize the variation of each trait.}

\item{trait.subset}{Character. Specific traits to measure. Defaults to NULL (all traits measured).}

\item{stom.val}{Numeric. Value of stomata (guard cells + pore). Defaults to 85.}

\item{subs.val}{Numeric. Value of subsidiary cells. Defaults to 170.}

\item{pave.val}{Numeric. Value of pavement cells. Defaults to 255}

\item{verbose}{Logical. Show progress of sampling? Defaults to TRUE.}
}
\value{
A data.frame with the differences in trait values (measured in standard
deviations if scaling is supplied) between each sampled cell patch and the whole image.
}
\description{
Measure epidermal traits from subsets of cells
}
