% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling_feature_extraction.r
\name{subsampling_feature_extraction}
\alias{subsampling_feature_extraction}
\title{Extract measurements from sampled_data.frame and polygons}
\usage{
subsampling_feature_extraction(sampled_polygons, sampled_data)
}
\arguments{
\item{sampled_polygons}{SpatialPolygonsDataFrame, containing the cell polygons.}

\item{sampled_data}{Data.frame, containing the measurements for all cells (some empty
columns because all cell types included in one file)}
}
\value{
A vector with whole-image values.
}
\description{
Extract measurements from sampled_data.frame and polygons
}
\details{
This function is designed to be used when the image has already
been processed and all cells have already been measured (e.g.
when evaluating sampling effort). For unmeasured images, see
\code{extract_image_features()}.
}
