% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling_trait_extraction.R
\name{subsampling_trait_extraction}
\alias{subsampling_trait_extraction}
\title{Extract measurements from sampled_data.frame and polygons}
\usage{
subsampling_trait_extraction(
  x,
  polygons,
  sample_index,
  trait.subset = NULL,
  stom.val = 85,
  subs.val = 170,
  pave.val = 255
)
}
\arguments{
\item{x}{List of the form output by \code{extract_epidermal_traits()} (see documentation of that function for details)}

\item{polygons}{sf object containing the cell polygons that were used to generate sample_index}

\item{sample_index}{Numeric vector of cells to be measured.}

\item{trait.subset}{Character. Specific traits to measure. Defaults to NULL (all traits measured).}

\item{stom.val}{Numeric. Value of stomata. Defaults to 85.}

\item{subs.val}{Numeric. Value of subsidiary cells. Defaults to 170.}

\item{pave.val}{Numeric. Value of pavement cells. Defaults to 255.}
}
\value{
A vector with whole-image values.
}
\description{
Extract measurements from sampled_data.frame and polygons
}
\details{
This function is designed to be used when the image has already
been processed and all cells have already been measured (e.g.
when evaluating sampling effort). For unmeasured images, see
\code{extract_epidermal_traits()}.
}
