% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{trait_key}
\alias{trait_key}
\title{Traits that can be measured}
\format{
A data frame with 60 observations and 5 variables
}
\usage{
trait_key
}
\description{
A dataset containing the names, cell types, categories and descriptions of all
traits that are extracted from images.
}
\keyword{datasets}
