% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_graph_shortest_paths.r
\name{cell_graph_shortest_paths}
\alias{cell_graph_shortest_paths}
\title{Find the shortest paths between each cell}
\usage{
cell_graph_shortest_paths(cells)
}
\arguments{
\item{cells}{SpatialPolygonsDataFrame.}
}
\value{
A matrix with the shortest paths between each cell
}
\description{
Find the shortest paths between each cell
}
