% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_extract_function.r
\name{extract_epidermal_traits}
\alias{extract_epidermal_traits}
\title{Extract cell measurements from an image}
\usage{
extract_epidermal_traits(
  cell.polygons,
  junction.points,
  cells.present = c("pavement", "stomata", "subsidiary"),
  cell.values,
  NA.value = NULL,
  stomatal.shape = TRUE,
  stomatal.arrangement = TRUE,
  sd.measures = TRUE,
  specific.inclusions = NULL,
  specific.exclusions = NULL,
  sd.as.percent.of.mean = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{junction.points}{Either a numeric matrix or SpatialPoints with the coordinates of the cell wall junctions (e.g. as output by \code{image_to_poly()})}

\item{cells.present}{Character. A vector containing the names of all the types of cells present. Defaults to \code{c("pavement", "stomata", "subsidiary")}}

\item{cell.values}{A vector containing the values of all the types of cells present (usually numeric), see details.}

\item{NA.value}{A vector of length 1, with the value for NA regions of the image (e.g. damaged or un-traced regions)}

\item{stomatal.shape}{Logical. Include measurements of stomatal shape?
Defaults to TRUE, see details.}

\item{stomatal.arrangement}{Logical. Include measurements of stomatal
arrangement? Defaults to TRUE, see details.}

\item{sd.measures}{Logical. Return standard deviation of variables?
Defaults to TRUE, see details.}

\item{specific.inclusions}{Character. Cell measurements to include;
see details for options.}

\item{specific.exclusions}{Character. Cell measurements to exclude;
see details for options.}

\item{sd.as.percent.of.mean}{Logical. Should variability measures be presented as % of the mean? Defaults to FALSE.}

\item{verbose}{Logical. Should progress output be printed in the console?
Defaults to FALSE.}

\item{cells}{SpatialPolygonsDataFrame. The image to be measured, converted to polygons (e.g. using \code{image_to_poly()})}
}
\value{
A list of length three, containing a vector with whole-image values, a
data.frame of pavement cell values and a data.frame of stomatal values.
}
\description{
Extract cell measurements from an image
}
\details{
This function contains the script for all measurements. Each type
of cell (e.g. pavement, stomate) should have a name and corresponding value
given in \code{cells.present} and \code{cell.values}, respectively. If
there are unnamed values in the image, these will be measured as
'othertype1', 'othertype2' and so forth. Only limited numbers of
measurements will be taken for these cell types (index, density, area,
spacing and distance to nearest neighbour). If only a subset of variables is desired,
this can can be set using the arguments \code{ stomatal.shape, stomatal.arrangement, pavement.cells, sd.measures,
specific.inclusions,  specific.exclusions}. For a full list of variables and
graphical descriptions of what they measure, see the vignette here. ((add link)).
}
