% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_edge_cells.r
\name{find_edge_cells}
\alias{find_edge_cells}
\title{Find edges of image, identify which get.edges.from are on the edge and which are}
\usage{
find_edge_cells(
  get.edges.from,
  find.edge.cells = NULL,
  edge.buffer.width = 2,
  rotated = NULL
)
}
\arguments{
\item{get.edges.from}{\code{sf} object. This object will be used to define the edges.}

\item{find.edge.cells}{\code{sf}object. This is the object that will be evaluated.
Defaults to \code{get.edges.from} (i.e. finds which cells of
\code{get.edges.from} lie on the edge of \code{get.edges.from} )}

\item{edge.buffer.width}{Numeric. The width (in pixels) that defines the edge. Defaults to 2.}

\item{rotated}{Numeric. The angle (in degrees) that the image has been rotated. Defaults to NULL.}
}
\value{
A list with two components.
 \describe{
\item{edge}{Numeric. Vector containing the indices of all edge get.edges.from.}
\item{lose}{Numeric. Vector containing the indices of all get.edges.from on the
            bottom or right-hand edge.}
}
}
\description{
Find edges of image, identify which get.edges.from are on the edge and which are
}
\details{
The \code{get.edges.from} parameter can be the output from
\code{image_to_poly()}.
}
