% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_junctions.r
\name{get_junctions}
\alias{get_junctions}
\title{Clean up junction point coordinates}
\usage{
get_junctions(wj, buffer.width)
}
\arguments{
\item{wj}{Two-column numeric matrix with x and y coordinates of junction points.}

\item{buffer.width}{Numeric, the distance within which two cell junctions are
merged. Suggested value: 5 pixels (or equivalent if image is scaled)}
}
\value{
\code{sf MULTIPOINT} object of the cleaned junction points
}
\description{
Clean up junction point coordinates
}
