% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_junctions.r
\name{get_junctions}
\alias{get_junctions}
\title{Clean up junction point coordinates}
\usage{
get_junctions(wj)
}
\arguments{
\item{wj}{Coordinates}
}
\value{
Some SpatialPoints
}
\description{
Clean up junction point coordinates
}
