% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_poly.r
\name{image_to_poly}
\alias{image_to_poly}
\title{Convert image to SpatialPolygons}
\usage{
image_to_poly(dir, file)
}
\arguments{
\item{dir}{String. Directory for images.}

\item{file}{String. File name.

#' @return A list with components:
\describe{
\item{cells}{sf MULTIPOLYGON. All cells, with values stored in a 'value' column.}
\item{junction_coords}{Dataframe with two columns: x and y coordinates of tri-cell junctions.}
}}
}
\description{
Convert image to SpatialPolygons
}
