% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_rows_stomata.r
\name{number_rows_stomata}
\alias{number_rows_stomata}
\title{Find the number of stomatal rows}
\usage{
number_rows_stomata(cells, stom.val)
}
\arguments{
\item{cells}{A `sf` object containing all cell polygons (i.e. not
just the stomatal polygons).}

\item{stom.val}{Numeric. The value (colour) of stomata.}
}
\value{
A list with components:
\describe{
\item{n.rows}{Numeric. The number of rows of stomata}
\item{row.bounds}{Numeric. The y coordinates defining the edge of each row}
\item{row.wiggliness}{Numeric. The width of the stomatal band; see Details}
\item{row.consistency}{Numeric. The number of stomata per row.}
\item{stomrow.density}{Numeric. The number of rows of stomata per unit
length (per pixel if unscaled)}
}
}
\description{
Find the number of stomatal rows
}
\details{
This function assumes that the pixels are coloured in the following order
(darkest to lightest): wall, stomate, subsidiary, pavement. To calculate
\code{row.wiggliness}, the major axis of the stomatal row is determined using
principal components, where PC1 is the main axis of the row and PC2 is the vertical
offset of each stomata from this axis. The value \code{row.wiggliness} is the
proportion of variation explained by PC2; a value close to 0 means that
stomata are perfectly aligned; a value close to 0.5 means that stomatal rows
are variable or wiggly in alignment. Values cannot exceed 0.5.

 @import stats
}
